Befebung von kleinen Bucks; 1. Mehrfaches login ist nun möglich ohne das die Alten blöcke weiter angezeigt werden 2.. Es werden nun die richtigen Blöcke Kontrolliert 3. Pasing des User unputs wurde verbessert.

+++ app/src/main/java/programm/Block.java
@@ -31,7 +31,7 @@ public class Block {
 
     public boolean checkAnswer(String userAnswer){
 
-        int answer = Integer.parseInt(userAnswer);
+        double answer = Double.parseDouble(userAnswer);
         return questions.get(currentQuestion).checkQuestion(answer);
     }
